requirements.txt
certifi
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.6.1
openai==1.109.1
pydantic==2.12.3
pydantic-core==2.41.4
PyYAML==6.0.2
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
volcengine-python-sdk[ark]==4.0.31
wheel==0.45.1
